reorganizing a code, moved some files from pages to components, added useContext with token from localStorage
