+++ back/models/Posts.js
@@ -1,7 +1,6 @@
 module.exports = (sequelize, DataTypes) => {
-
+    
     const Posts = sequelize.define("Posts", {
-        
         postText: {
             type: DataTypes.STRING,
             allowNull: false,
@@ -10,17 +9,14 @@ module.exports = (sequelize, DataTypes) => {
             type: DataTypes.STRING,
             allowNull: true,
         }
-    })
+    });
 
     Posts.associate = (models) => {
         Posts.hasMany(models.Likes, {
             onDelete: 'cascade',
-        })
+        }),
 
         Posts.belongsTo(models.Users)
     }
-
-
-
     return Posts;
 }
