+++ front/src/components/Dropdown.js
@@ -8,12 +8,12 @@ export default function Dropdown() {
         localStorage.removeItem('accessToken');
         navigate('/auth/login', { replace: true });
     }
-  return (
+    return (
         <ul className="dropdown-menu" aria-labelledby="dropdownMenuButton1">
             <li>
                 <Link className='dropdown-item' to={'/auth/profile'}>Profile</Link>
             </li>
             <li><button onClick={() => logOut()} type='button' className="dropdown-item">Log out</button></li>
         </ul>
-  )
+    )
 }
