--- front/src/pages/auth/Auth.css
@@ -1,14 +1,0 @@
-.btn-primaire { 
-    background-color: #FF2D01 !important; 
-    color: white;
-    font-weight: 700;
-}
-.btn-primaire:hover, .dropdown-item:hover { 
-    background-color: #FFD7D7 !important;
-    color: #FF2D01 !important; 
-    outline: none !important;
-}
-
-.btn-primaire:hover a {
-    color: #FF2D01 !important;
-}
