+++ front/src/pages/auth/SignUp.js
@@ -1,10 +1,9 @@
-import {React, useState} from 'react'
+import {React, useState} from 'react';
 import axios from 'axios';
-import {useNavigate} from 'react-router-dom'
+import {useNavigate} from 'react-router-dom';
+import Header from '../../components/header/Header';
 
 
-import Header from '../header/Header'
-import './Auth.css'
 
 export default function SignUp() {
