--- front/src/pages/home/Home.css
@@ -1,6 +1,0 @@
-.red {
-    color: red;
-}
-.black {
-    color: black;
-}
