+++ front/src/pages/home/Home.js
@@ -1,13 +1,11 @@
-import './Home.css'
-import Header from '../header/Header';
-
+import Header from '../../components/header/Header';
 import AllPosts from '../../components/AllPosts';
 
 export default function Home() {
   return (
-    <div className='bg-light '>
-      <Header/>
-      <AllPosts/>
-    </div>
+      <div className='bg-light '>
+        <Header/>
+        <AllPosts/>
+      </div>
   )
 }
